requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
Pillow>=9.5.0
pydicom>=2.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# HAI-DEF Models
transformers>=4.35.0
bitsandbytes>=0.41.0
accelerate>=0.25.0

# Dimensionality Reduction
umap-learn>=0.5.3

# Progress and Logging
tqdm>=4.66.0
tensorboard>=2.14.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.6.0

# Development
ipython>=8.0.0
jupyter>=1.0.0
